// src/app/Life/services/quotation.interface.ts
export interface IDiscountCode {
  description: string;
  code: string;
}
export interface IPremiumFrequency {
    id:          number;
    premFreq:    string;
    description: string;
    noPayYear:   number;
    createdBy:   string;
    frequency:   string;
}
export interface IPremiumPaymentMethod {
  description: string;
  code: string;
}
// ...
export interface IBenefitPaymentFreq {
  paymentFrequency: string;
}
// ...
export interface ICoverInfo {
    id:                                 number;
    productCode:                        string;
    subgroupNo:                         string;
    availablePostIssue:                 boolean;
    sumAssuredEqualBaseCoverSumAssured: boolean;
    blankSumAssured:                    boolean;
    mandatory:                          boolean;
    company:                            string;
    ageAssuredLimit:                    string;
    coverCode:                          string;
    coolOffPeriod:                      number;
    fundingOption:                      string;
    maxNo:                              string;
    description:                        string;
    base:                               string;
    code:                               string;
    coverSumAssured:                    string;
    riderCoverCode:                     string;
}
export enum CoverScreenNames {
    BL = "BUILDING",
    DR = "DEFENSE_RECOURSE", 
    LIL = "LIFE_INSURANCE_LIFE", 
    AL = "ACCESSORIES_RISK", // /
    TP = "THIRDPARTY", 
    PC = "PROPERTY_CONTENT", 
    SAV = "SAVINGS", 
    MH = "MARINE_HULL", 
    MC = "MARINE_CARGO",
    LA = "LIFE_ANNUITY",
    AC = "ANNUITIES_CERTAIN",
    ILR = "INDIVIDUAL_LIFE_RISK",
    ICL = "INDIVIDUAL_CREDIT_LIFE",
    IS = "INDIVIDUAL_SAVINGS",// /
    END = "ENDOWMENT",
    GLR = "GROUP_CREDIT_LIFE",// /
    BO = "BOND"
}

export type CoverCode = 'AC' | 'LA' | 'ILR' | 'ICR'| 'IS' | 'END' | 'GLR' | 'GS' | 'TP' | 'DR' | 'AL' | 'LIL' | 'BL' | 'PC' | 'SAV' | 'MH' | 'MC' | 'ICL' | 'BO'

export interface ICoverScreen {
  description: string;
  coverScreen: CoverCode;
}

export interface ICoversDetails {
  info: ICoverInfo;
  screen: {
    code: CoverCode;
    name: CoverScreenNames;
  };
}